refactor: Make FadingScrollbar a react component (#14279)
