+++ src/script/components/Modals/GroupCreation/GroupCreationModal.tsx
@@ -28,6 +28,7 @@ import {container} from 'tsyringe';
 import {Button, ButtonVariant, Select} from '@wireapp/react-ui-kit';
 import {WebAppEvents} from '@wireapp/webapp-events';
 
+import {FadingScrollbar} from 'Components/FadingScrollbar';
 import {Icon} from 'Components/Icon';
 import {ModalComponent} from 'Components/ModalComponent';
 import {SearchInput} from 'Components/SearchInput';
@@ -36,7 +37,6 @@ import {BaseToggle} from 'Components/toggle/BaseToggle';
 import {InfoToggle} from 'Components/toggle/InfoToggle';
 import {UserSearchableList} from 'Components/UserSearchableList';
 import {useKoSubscribableChildren} from 'Util/ComponentUtil';
-import {initFadingScrollbar} from 'Util/DOM/fadingScrollbar';
 import {handleEnterDown, offEscKey, onEscKey} from 'Util/KeyboardUtil';
 import {t} from 'Util/LocalizerUtil';
 import {getLogger} from 'Util/Logger';
@@ -354,7 +354,7 @@ const GroupCreationModal: React.FC<GroupCreationModalProps> = ({
         )}
 
         {stateIsParticipants && (
-          <div className="group-creation__list" ref={initFadingScrollbar}>
+          <FadingScrollbar className="group-creation__list">
             {contacts.length > 0 && (
               <UserSearchableList
                 users={contacts}
@@ -367,7 +367,7 @@ const GroupCreationModal: React.FC<GroupCreationModalProps> = ({
                 noUnderline
               />
             )}
-          </div>
+          </FadingScrollbar>
         )}
         {/* eslint jsx-a11y/no-autofocus : "off" */}
         {stateIsPreferences && (
