+++ src/script/page/RightSidebar/GroupParticipantService/GroupParticipantService.tsx
@@ -19,10 +19,10 @@
 
 import {FC, useEffect} from 'react';
 
+import {FadingScrollbar} from 'Components/FadingScrollbar';
 import {Icon} from 'Components/Icon';
 import {ServiceDetails} from 'Components/panel/ServiceDetails';
 import {useKoSubscribableChildren} from 'Util/ComponentUtil';
-import {initFadingScrollbar} from 'Util/DOM/fadingScrollbar';
 import {handleKeyDown} from 'Util/KeyboardUtil';
 import {t} from 'Util/LocalizerUtil';
 import {matchQualifiedIds} from 'Util/QualifiedId';
@@ -97,7 +97,7 @@ const GroupParticipantService: FC<GroupParticipantServiceProps> = ({
     <div id="group-participant-service" className="panel__page group-participant">
       <PanelHeader onGoBack={onBack} goBackUie="go-back-group-participant" onClose={onClose} />
 
-      <div className="panel__content panel__content--fill" ref={initFadingScrollbar}>
+      <FadingScrollbar className="panel__content panel__content--fill">
         <ServiceDetails service={serviceEntity} />
 
         {showActions && canChatWithServices() && (
@@ -133,7 +133,7 @@ const GroupParticipantService: FC<GroupParticipantServiceProps> = ({
             <div className="panel__action-item__text">{t('groupParticipantActionRemove')}</div>
           </div>
         )}
-      </div>
+      </FadingScrollbar>
 
       {isAddMode && (
         <div className="panel__footer">
