+++ src/script/page/RightSidebar/TimedMessages/TimedMessages.tsx
@@ -19,8 +19,8 @@
 
 import {FC, useEffect, useState} from 'react';
 
+import {FadingScrollbar} from 'Components/FadingScrollbar';
 import {useKoSubscribableChildren} from 'Util/ComponentUtil';
-import {initFadingScrollbar} from 'Util/DOM/fadingScrollbar';
 import {t} from 'Util/LocalizerUtil';
 import {formatDuration} from 'Util/TimeUtil';
 
@@ -91,7 +91,7 @@ const TimedMessages: FC<TimedMessagesPanelProps> = ({activeConversation, onClose
         goBackUie="go-back-timed-messages-options"
       />
 
-      <div ref={initFadingScrollbar} className="panel__content">
+      <FadingScrollbar className="panel__content">
         {messageTimes.map(({text, isCustom, value}) => (
           <label
             key={value}
@@ -111,7 +111,7 @@ const TimedMessages: FC<TimedMessagesPanelProps> = ({activeConversation, onClose
         ))}
 
         <p className="panel__info-text timed-messages__disclaimer">{t('timedMessageDisclaimer')}</p>
-      </div>
+      </FadingScrollbar>
     </div>
   );
 };
